requirements.txt
openai
anthropic
streamlit
streamlit-ace
streamlit-modal
python-dotenv
duckduckgo-search
requests
beautifulsoup4
arxiv
streamlit-option-menu
pypdf
youtube-transcript-api
clipboard
googlesearch-python
